pythonProject/requirements.txt
mysql-connector-python
flask
flask-jwt-extended
requests
python-dotenv
mysql
werkzeug
PyJWT
itsdangerous